basic env config

+++ src/categories/categories.service.ts
@@ -4,6 +4,7 @@ import { UpdateCategoryDto } from './dto/update-category.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DeleteResult, Repository, UpdateResult } from 'typeorm';
 import { Category } from './entities/category.entity';
+import { IdNotFoundException } from '../exceptions/id-not-found-exception';
 
 @Injectable()
 export class CategoriesService {
@@ -11,24 +12,41 @@ export class CategoriesService {
   constructor(@InjectRepository(Category) private readonly categoryRepository: Repository<Category>) {
   }
 
-  create(createCategoryDto: CreateCategoryDto): Promise<Category> {
-    return this.categoryRepository.save(createCategoryDto);
+  async create(createCategoryDto: CreateCategoryDto): Promise<Category> {
+    return this.categoryRepository.save(this.getCategory(new Category(), createCategoryDto));
   }
 
   findAll(): Promise<Category[]> {
     return this.categoryRepository.find();
   }
 
-  findOne(id: number): Promise<Category> {
-    return this.categoryRepository.findOne(id);
+  async findById(id: number): Promise<Category> {
+    return await this.getCategoryById(id);
   }
 
-  update(id: number, updateCategoryDto: UpdateCategoryDto): Promise<UpdateResult> {
-    return this.categoryRepository.update(id, updateCategoryDto);
+  async update(id: number, updateCategoryDto: UpdateCategoryDto): Promise<UpdateResult> {
+    if (!!await this.getCategoryById(id)) {
+      return this.categoryRepository.update(id, this.getCategory(new Category, updateCategoryDto));
+    }
   }
 
-  remove(id: number): Promise<DeleteResult> {
-    return this.categoryRepository.delete(id);
+  async remove(id: number): Promise<DeleteResult> {
+    if (!!await this.getCategoryById(id)) {
+      return this.categoryRepository.delete(id);
+    }
+  }
+
+  getCategory(currentCategory: Category, category: CreateCategoryDto | UpdateCategoryDto): Category {
+    currentCategory.category = category.category;
+    return currentCategory;
+  }
+
+  async getCategoryById(id: number): Promise<Category> {
+    try {
+      return await this.categoryRepository.findOneOrFail(id);
+    } catch (err) {
+      throw new IdNotFoundException({ id });
+    }
   }
 
 }
